requirements.txt
streamlit>=1.33.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
pypdf>=4.2.0
numpy>=1.23.0
pandas>=2.0.0
pyyaml>=6.0.1
openai>=1.40.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
